# Hash first in _make_input_signature_hashable