@jonathanchu33 2020-08-18
Benchmarks over 3 batches of 10 trials each

Minimum
defun_matmul_2_by_2_CPU: 164 -> 148 us
defun_matmul_2_by_2_CPU_async: 83.6 -> 80.5 us
defun_matmul_2_by_2_with_signature_CPU: 219 -> 195 us

Mean
defun_matmul_2_by_2_CPU: 175 -> 156 us
defun_matmul_2_by_2_CPU_async: 86 -> 82.5 us
defun_matmul_2_by_2_with_signature_CPU: 236 -> 213 us

Median
defun_matmul_2_by_2_CPU: 172-> 156 us
defun_matmul_2_by_2_CPU_async: 86 -> 82.5 us
defun_matmul_2_by_2_with_signature_CPU: 227 -> 204 us